+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix: Decorates the main Doctrine service IDs, not the aliases.
 * fix: Use `callable|array` typehint when decorating event dispatchers. See https://github.com/symfony/symfony/issues/48130
 * test: refactor tests
+* feat: supports multiple entity managers.
 
 ## 1.2.5